refactor: broken code into components,  fix: cart items count
