+++ src/components/AddToCartButton/AddToCartButton.js
@@ -1,4 +1,4 @@
-import { useCart } from "../contexts";
+import { useCart } from "../../contexts";
 
 export const AddToCartButtton = ({ item }) => {
   const { dispatch } = useCart();
