+++ src/components/Products/Products.js
@@ -1,17 +1,10 @@
-import { useEffect, useReducer, useState } from "react";
-import { data } from "../data";
-import { useCart } from "../contexts/cart-context";
-import { UpdateCartButton } from "./UpdateCartButton";
-import { AddToCartButtton } from "./AddToCartButtton";
-import { useWishlist } from "../contexts/wishlist-context";
-import { AddToWishlistButton } from "./AddToWishlistButton";
-import { RemoveFromWishlistButton } from "./RemoveFromWishlistButton";
-import { useFilter } from "../contexts/sort-and-filter-context";
+import { useEffect, useState } from "react";
+import { data } from "../../data";
+import { useFilter } from "../../contexts";
+import { ProductCard } from "../ProductCard/ProductCard";
 
 export const Products = () => {
   const [products, setProducts] = useState([]);
-  const { cart } = useCart();
-  const { wishlist } = useWishlist();
   const {
     sortBy,
     showFastDeliveryOnly,
@@ -21,7 +14,6 @@ export const Products = () => {
     getSortedData,
     getFilteredData
   } = useFilter();
-  let currCartItem, currWishlistItem;
 
   useEffect(() => setProducts(data), [products]);
 
@@ -94,7 +86,7 @@ export const Products = () => {
           Price Range : 0 to {rangeValue}
         </label>
 
-        <div class="button-container">
+        <div className="button-container">
           <button
             onClick={() => dispatch({ type: "CLEAR_ALL_FILTERS" })}
             className="btn btn-secondary btn-style"
@@ -106,47 +98,7 @@ export const Products = () => {
 
       <div className="grid-row grid-wrapper">
         {filteredProducts.map((productItem) => {
-          return (
-            <div key={productItem.id}>
-              <div className="vertical-card">
-                <div className="thumbnail">
-                  <img
-                    className="img-responsive-vertical"
-                    src={productItem.image}
-                    alt=""
-                  />
-                  {wishlist.find((wishlistItem) => {
-                    if (wishlistItem.id === productItem.id) {
-                      currWishlistItem = productItem;
-                      return wishlistItem;
-                    }
-                    return null;
-                  }) ? (
-                    <RemoveFromWishlistButton item={currWishlistItem} />
-                  ) : (
-                    <AddToWishlistButton item={productItem} />
-                  )}
-                  {/*  */}
-                </div>
-                <div className="product-description">
-                  <h3> {productItem.name} </h3>
-                  <p> Price: Rs {productItem.price} </p>
-
-                  {cart.find((cartItem) => {
-                    if (cartItem.id === productItem.id) {
-                      currCartItem = cartItem;
-                      return cartItem;
-                    }
-                    return null;
-                  }) ? (
-                    <UpdateCartButton item={currCartItem} />
-                  ) : (
-                    <AddToCartButtton item={productItem} />
-                  )}
-                </div>
-              </div>
-            </div>
-          );
+          return <ProductCard productItem={productItem} />;
         })}
       </div>
     </>
