+++ src/components/index.js
@@ -1,2 +1,5 @@
-export { cart } from "./Cart";
-export { wishlist } from "./Wishlist";
+export { UpdateCartButton } from "./UpdateCartButton/UpdateCartButton";
+export { Cart } from "./Cart/Cart";
+export { Wishlist } from "./Wishlist/Wishlist";
+export { Products } from "./Products/Products";
+export { Navbar } from "./Navbar/Navbar";
